created database and posts table and created a post.
connected that databse with our laravel app.

slug is one of the columns that has the messagge 'my-first-post' that we pass that in the url at the end.

$post->body :
we do this because is an object and we wanna access the body column that has text that is what we will see on our page:

http://127.0.0.1:8000/posts/my-first-post#