Changed to sass in the webpack.mix.js package to get support for sass.
In webpack we can changed to for example .react() to get support for react.

added some things in package.json and run npm install.

npm run dev: so we run all those dependencies.

npm run watch: to watch all the changes that we have done in app.scss and add those changes automatically in app.css in the public folder.