in layout changed the request path to articles and changed the css links to access those css files.
in web added a new Route to the articles depending on which id we pass we show a different article.

created articles controller with its show function to find that article depending on the id wqe pass and rendering it on the page by using view() and accesing the show file and passing that article.

created the show page for the articles in a new folder called articles.

added an anchor tag that takes us to the specific article that we clicked on in teh about page.