added the articles page where we display all the articles.
added new route to /articles where we use the function index() in the ArticlesController.

created index() function in ArticlesController where we dispplay all the articles.

added /articles to the Request::path articles.

created the index file in the articles folder whre we loop through all the articles and display them.